reordered position, velocity, acceleration assignments to fix conservation of energy

+++ Engine/Vec2.cpp
@@ -6,47 +6,38 @@ Vec2 Vec2::Unit()
 {
 	return Vec2( x / Mag(), y / Mag());
 }
-
 Vec2 Vec2::Cross( Vec2 vector)
 {
 	// cross product calculation
 
 	return Vec2(0.0f, 0.0f);
 }
-
 float Vec2::Dot(Vec2 vector)
 {
 	// dot product calculation
 
 	return 0.0f;
 }
-
 Vec2 Vec2::Subtract(Vec2 vector)
 {
 	return Vec2(x - vector.GetX(), y - vector.GetY());
 }
-
 Vec2 Vec2::Add(Vec2 vector)
 {
 	return Vec2(x + vector.GetX(), y + vector.GetY());
 }
-
 Vec2 Vec2::Scale(float scalar)
 {
 	return Vec2(x * scalar, y * scalar);
 }
-
 float Vec2::Mag()
 {
 	return sqrtf(x * x + y * y);
 }
-
 float Vec2::Mag2()
 {
 	return x * x + y * y;
 }
-
-
 
 
 // interface functions
@@ -54,12 +45,10 @@ void Vec2::InvertX()
 {
 	x = -x;
 }
-
 void Vec2::InvertY()
 {
 	y = -y;
 }
-
 
 
 // Draw function
@@ -73,16 +62,16 @@ void Vec2::DrawVec(Graphics& gfx, Vec2 start)
 	{
 		for (int x = 0; x <= (int)unitx; x++)
 		{
-			int pixelx = i * unitx + x + startx;
-			int pixely = i * unity + starty;
+			int pixelx = i * (int)unitx + x + (int)startx;
+			int pixely = i * (int)unity + (int)starty;
 			if (pixelx < gfx.ScreenWidth && pixely < gfx.ScreenHeight && pixelx > 0 && pixely > 0) {
 				gfx.PutPixel(pixelx, pixely, 255, 255, 255);
 			}
 		}
 		for (int y = 0; y <= (int)unity; y++)
 		{
-			int pixelx = i * unitx + startx;
-			int pixely = i * unity + y + starty;
+			int pixelx = i *(int) unitx + (int)startx;
+			int pixely = i * (int)unity + y + (int)starty;
 			if (pixelx < gfx.ScreenWidth && pixely < gfx.ScreenHeight && pixelx > 0 && pixely > 0) {
 				gfx.PutPixel(pixelx, pixely, 255, 255, 255);
 			}
@@ -91,13 +80,11 @@ void Vec2::DrawVec(Graphics& gfx, Vec2 start)
 }
 
 
-
 // coordinate getter functions
 float Vec2::GetX()
 {
 	return x;
 }
-
 float Vec2::GetY()
 {
 	return y;
